+++ src/Dofus/Types/GameRolePlayActorInformations.cs
@@ -5,7 +5,7 @@ namespace Dofus.Types
 {
     public class GameRolePlayActorInformations : GameContextActorInformations, INetworkMessage
     {
-        internal new static int MessageId => 5720;
+        internal new static ushort MessageId => 5720;
 
         public new void Serialize(DofusBinaryWriter writer)
         {
