+++ src/Dofus/Types/GameRolePlayCharacterInformations.cs
@@ -5,7 +5,7 @@ namespace Dofus.Types
 {
     public class GameRolePlayCharacterInformations : GameRolePlayHumanoidInformations, INetworkMessage
     {
-        internal new static int MessageId => 8148;
+        internal new static ushort MessageId => 8148;
 
         public ActorAlignmentInformations AlignmentInfos { get; private set; } = new();
