+++ src/Dofus/Types/HumanOptionEmote.cs
@@ -5,7 +5,7 @@ namespace Dofus.Types
 {
     public class HumanOptionEmote : HumanOption, INetworkMessage
     {
-        internal new static int MessageId => 5644;
+        internal new static ushort MessageId => 5644;
 
         public int EmoteId { get; private set; }
         public double StartTime { get; private set; }
